+++ packages/contracts/deploy/update/to_v1.3.0/10_Multisig_Plugin.ts
@@ -3,8 +3,12 @@ import {HardhatRuntimeEnvironment} from 'hardhat/types';
 import {PluginRepo__factory} from '../../../typechain';
 import {getContractAddress, uploadToIPFS} from '../../helpers';
 
+import multisigSetupArtifact from '../../../artifacts/src/plugins/governance/multisig/MultisigSetup.sol/MultisigSetup.json';
+
 import multisigReleaseMetadata from '../../../src/plugins/governance/multisig/release-metadata.json';
 import multisigBuildMetadata from '../../../src/plugins/governance/multisig/build-metadata.json';
+
+const TARGET_RELEASE = 1;
 
 const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   console.log('\nUpdate Multisig Plugin');
@@ -13,6 +17,7 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   const [deployer] = await ethers.getSigners();
 
   const deployResult = await deploy('MultisigSetup', {
+    contract: multisigSetupArtifact,
     from: deployer.address,
     args: [],
     log: true,
@@ -44,7 +49,7 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
     const tx = await multisigRepo
       .connect(deployer)
       .createVersion(
-        1,
+        TARGET_RELEASE,
         deployResult.address,
         ethers.utils.toUtf8Bytes(`ipfs://${multisigBuildCIDPath}`),
         ethers.utils.toUtf8Bytes(`ipfs://${multisigReleaseCIDPath}`)
@@ -57,7 +62,7 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
   const tx = await multisigRepo
     .connect(deployer)
     .populateTransaction.createVersion(
-      1,
+      TARGET_RELEASE,
       deployResult.address,
       ethers.utils.toUtf8Bytes(`ipfs://${multisigBuildCIDPath}`),
       ethers.utils.toUtf8Bytes(`ipfs://${multisigReleaseCIDPath}`)