+++ packages/contracts/CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Inherit `ProtocolVersion` and `ERC165` in `DAOFactory`.
+- Inherit `ProtocolVersion` in `DAO`.
 - Added a `nonReentrant` modifier to the `execute` function in the `DAO` contract.
 - Added `allowFailureMap` to `IDAO.Executed` event.